pyproject.toml
[project]
name = "mmm-attack"
version = "0.1.0"
description = "Multi-Modal Multi-turn Jailbreak Attack System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.40.0; sys_platform != 'darwin'",
    "auto-gptq>=0.4.0; sys_platform != 'darwin'",
    "autoawq>=0.1.0; sys_platform != 'darwin'",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.40",
    "anthropic>=0.25.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "google-api-core>=2.11.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[project.scripts]
mmm-attack = "mmm_attack.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mmm_attack"]